+++ packages/matter.js/src/behavior/cluster/ClusterBehaviorUtil.ts
@@ -9,7 +9,7 @@ import { ClusterType } from "../../cluster/ClusterType.js";
 import { ImplementationError } from "../../common/MatterError.js";
 import { AttributeModel, ClusterModel, ElementTag, FeatureSet, Matter, Metatype } from "../../model/index.js";
 import { GeneratedClass } from "../../util/GeneratedClass.js";
-import { Observable } from "../../util/Observable.js";
+import { AsyncObservable } from "../../util/Observable.js";
 import { camelize } from "../../util/String.js";
 import { Behavior } from "../Behavior.js";
 import { DerivedState } from "../state/StateType.js";
@@ -185,7 +185,7 @@ function createDerivedEvents(cluster: ClusterType, base: Behavior.Type, stateNam
 
     // Add events for mandatory attributes that are not present in the base class
     for (const attrName of stateNames) {
-        const name = `${attrName}$Change`;
+        const name = `${attrName}$Changed`;
         if (baseInstance[name] === undefined) {
             names.add(name);
         }
@@ -199,7 +199,7 @@ function createDerivedEvents(cluster: ClusterType, base: Behavior.Type, stateNam
 
         initialize() {
             for (const name of names) {
-                (this as any)[name] = Observable();
+                (this as any)[name] = AsyncObservable();
             }
         },
     });