+++ CHANGELOG.md
@@ -20,11 +20,13 @@ The main work (all changes without a GitHub username in brackets in the below li
   * Enhancement: Adds Certificate validation and cryptographic verification during commissioning and CASE session establishment
   * Enhancement: Adds additional logging information for PASE and CASE to better understand errors without debug logging
   * Fix: Corrects returned errors for two commands on OperationalCredentials cluster 
-* matter.js New API code flows: 
+* matter.js New API code flows:
+  * Breaking: The name of the *$Change Events for attributes and such are changed to *$Changed . Please adjust your code!
   * Enhancement: Optimized constraint validations and conformance error messages
   * Enhancement: Conditionally enables the ReachableChanged event on the Root Endpoint BasicInformation cluster if the reachable attribute is defined in the defaults
   * Enhancement: Allow to register events directly when initializing endpoints like in legacy API
-  * Enhancement: Allows for cluster implementations to dynamically add/enable state attributes and events 
+  * Enhancement: Allows for cluster implementations to dynamically add/enable state attributes and events
+  * Enhancement: Allows "$Changed" event handlers to be async
   * Fix: Fixes some issues around event handling in the new API and makes sure events are not de-registered on factory resets
   * Fix: Corrects the returned status error code when an Enum value is set to an invalid value
 * Chip testing: